fix: HttpError and retry err parsing

+++ crates/transport-http/src/hyper_transport.rs
@@ -88,7 +88,7 @@ where
 
                 if status != hyper::StatusCode::OK {
                     return Err(TransportErrorKind::http_error(
-                        status.as_u16() as i64,
+                        status.as_u16(),
                         String::from_utf8_lossy(&body).to_string(),
                     ));
                 }
