+++ crates/transport-http/src/reqwest_transport.rs
@@ -64,7 +64,7 @@ impl Http<Client> {
 
                 if status != reqwest::StatusCode::OK {
                     return Err(TransportErrorKind::http_error(
-                        status.as_u16() as i64,
+                        status.as_u16(),
                         String::from_utf8_lossy(&body).to_string(),
                     ));
                 }
